Updates   * Added more style injector wrappers around primitive compoents     * Updated input, image, touchable     * Cleaned up view and text components
