+++ repos/keg-components/configs/aliases.json
@@ -4,18 +4,19 @@
   "KegFilePicker": "src/components/filePicker/filePicker${platform}.js",
   "KegForm": "src/components/form/form/form.js",
   "KegImg": "src/components/image/image.js",
+  "KegInternalImage": "src/components/image/kegImage${platform}.js",
   "KegIcon": "src/components/icon/icon.js",
   "KegIcons": "src/assets/icons",
   "KegIndicator": "src/components/indicator/indicator.js",
-  "KegInput": "src/components/form/input/input.js",
+  "KegInput": "src/components/form/input/input${platform}.js",
   "KegLink": "src/components/link/link.js",
   "KegOption": "src/components/form/option/option${platform}.js",
   "KegRadio": "src/components/form/radio/radio${platform}.js",
   "KegSelect": "src/components/form/select/select.js",
   "KegSlider": "src/components/form/slider/slider${platform}.js",
   "KegSwitch": "src/components/form/switch/switch.js",
-  "KegText": "src/components/typography/kegtext.js",
-  "KegView": "src/components/view/view.js",
+  "KegText": "src/components/typography/kegtext${platform}.js",
+  "KegView": "src/components/view/view${platform}.js",
   "KegGetPlatform": "src/utils/helpers/getPlatform.js",
   "KegAnimateHook": "src/hooks/useAnimate${platform}.js",
   "KegAssets": "src/assets/index.js",
@@ -28,7 +29,5 @@
   "KegClassName": "src/hooks/useClassName${platform}.js",
   "KegTypeAsClass": "src/hooks/useThemeTypeAsClass${platform}.js",
   "KegInjector": "src/components/styleInjector/styleInjector.js",
-  "KegStyleTag": "src/components/styleInjector/useStyleTag.js",
-  "KegInjectedText": "src/components/typography/kegtext.injected.js",
-  "KegInjectedView": "src/components/view/view.injected.js"
+  "KegStyleTag": "src/components/styleInjector/useStyleTag.js"
 }
