+++ repos/re-theme/src/styleInjector/styleInjector.native.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+/**
+ * Placeholder for native compoennts. Just returns the passed in component
+ * @param {Object} Component - React Component to wrap
+ * @param {Object} config - Settings for the Hoc
+ * 
+ * @returns {Component} - The passed in component
+ */
+export const StyleInjector = Component => Component
+
